pyproject.toml
[project]
name = "llearn"
version = "0.1.0"
description = ""
authors = [
    {name = "kikiru328",email = "kikiru328@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "redis[async] (>=6.4.0,<7.0.0)",
    "hiredis (>=3.2.1,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
faker = "^37.4.2"
freezegun = "^1.5.4"
pytest-mock = "^3.14.1"
pytest-asyncio = "^1.1.0"
greenlet = "^3.2.3"

